Display iterated pokemon
Move opening and closing tags outside of IIFE so they can be accessed. Delete forEach() function within IIFE, this is not needed as forEach() is native to Javascript. Replace for loop with forEach() loop, nest conditions within forEach() loop.

Update the language of the conditions within the forEach() loop. For example, change repository[i].height to pokemon.height.